MO-10870 Alter submodule URLs to pull from code.qt.io
Changed the URLs of the standard/unaltered Qt submodules
to point to code.qt.io which is the main code repository
for the Qt project.

The previous relative URL addresses used the GitHub mirror
of the Qt project which is not always in sync with the
code.qt.io master.