MO-10870: Set qtbase to new submodule branch and alter submodule URLs to pull from code.qt.io
Update qtbase branch to new merged bfx/5.15.2.

The copied David Barker's changes from bfx/15.1 branch:
Changed the URLs of the standard/unaltered Qt submodules
to point to code.qt.io which is the main code repository
for the Qt project.

The previous relative URL addresses used the GitHub mirror
of the Qt project which is not always in sync with the
code.qt.io master.